+++ src/main/java/org/apptopia/waterwayfreightsystem/api/api/application/CargoService.java
@@ -6,7 +6,7 @@ import org.apptopia.waterwayfreightsystem.api.api.application.usecases.cargo.Raw
 import org.apptopia.waterwayfreightsystem.api.api.application.usecases.cargo.RawCargoOutput;
 
 public interface CargoService {
-	List<RawCargoOutput> getCargosOfCustomer(Integer idUser);
+	List<RawCargoOutput> getCargosOfCustomer(Long idUser);
 	List<RawCargoOutput> findAllCargos();
 	RawCargoOutput newCargo(RawCargoInput rawCargoInput);
 	RawCargoOutput updateCargo(RawCargoInput rawCargoInput);
