+++ src/main/java/org/apptopia/waterwayfreightsystem/api/api/application/usecases/orders/RawOrdersMapper.java
@@ -29,7 +29,7 @@ public interface RawOrdersMapper {
 		return ship.getIdShip();
 	}
 	
-	default Account toAccount(Integer idUser) {
+	default Account toAccount(Long idUser) {
 		
 		if(idUser == null || idUser == 0) {
 			return null;
@@ -37,7 +37,7 @@ public interface RawOrdersMapper {
 		return new Account(idUser);
 	}
 	
-	default Integer fromAccount(Account account) {
+	default Long fromAccount(Account account) {
 		
 		if(account == null) {
 			return null;
